chore: extract withResolver in a utility
