+++ test/utils/promise.js
@@ -1,0 +1,10 @@
+const withResolvers = function () {
+  let promiseResolve, promiseReject
+  const promise = new Promise((resolve, reject) => {
+    promiseResolve = resolve
+    promiseReject = reject
+  })
+  return { promise, resolve: promiseResolve, reject: promiseReject }
+}
+
+module.exports = { withResolvers }
